pyproject.toml
[tool.poetry]
name = "auction"
version = "0.1.0"
description = ""
authors = ["Dmitry Bakunovich <dv.bakunovich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1"
psycopg2 = "^2.9.1"
flake8 = "^3.9.2"
djangorestframework = "3.12.2"
Pillow = "^8.3.2"
django-filter = "^21.1"
djangorestframework-simplejwt = "^4.8.0"
celery = "^5.1.2"
redis = "^3.5.3"
ipython = "^7.28.0"
django-cors-headers = "^3.10.0"
flower = "^1.0.0"
channels = "^3.0.4"
channels-redis = "^3.3.1"
pytest = "^6.2.5"
pytest-django = "^4.4.0"
pytest-factoryboy = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
